config center, daily

+++ services/shared/db/model/common.go
@@ -9,7 +9,7 @@ type Common struct {
 	ID             string        `pg:",pk"`
 	CreatedTime    time.Duration `pg:",use_zero,notnull"`
 	UpdateTime     time.Duration `pg:",use_zero,notnull"`
-	OptimisticLock int           `pg:",use_zero,notnull"`
+	OptimisticLock uint64        `pg:",use_zero,notnull"`
 }
 
 func NewCommon() Common {
